refactor(#15): util 관련 파일 경로 수정
