+++ application/auth/src/main/java/io/jus/hopegaarden/config/security/filter/JwtAuthenticationFilter.java
@@ -3,21 +3,18 @@ package io.jus.hopegaarden.config.security.filter;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.SignatureException;
-import io.jus.hopegaarden.domain.define.member.Member;
-import io.jus.hopegaarden.domain.define.member.constant.MemberRole;
 import io.jus.hopegaarden.domain.define.token.jwt.repository.TokenRepository;
 import io.jus.hopegaarden.exception.ErrorCode;
 import io.jus.hopegaarden.exception.exceptions.auth.*;
 import io.jus.hopegaarden.exception.util.ErrorUtil;
 import io.jus.hopegaarden.sevice.jwt.JwtTokenProvider;
-import io.jus.hopegaarden.utils.AuthenticationExtractor;
+import auth.AuthenticationExtractor;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -28,8 +25,6 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
-import java.util.Arrays;
-import java.util.List;
 import java.util.Optional;
 
 @Slf4j
